# Improve initial build performance by reducing calls to `synchronizeHostData`