--- a/test/declarationWatch/expectedOutput/patch0/output.txt
+++ b/test/declarationWatch/expectedOutput/patch0/output.txt
@@ -1,6 +1,6 @@
-    Asset     Size  Chunks             Chunk Names
-bundle.js  1.58 kB       0  [emitted]  main
-chunk    {0} bundle.js (main) 112 bytes [rendered]
+    Asset     Size  Chunks       Chunk Names
+bundle.js  1.58 kB       0       main
+chunk    {0} bundle.js (main) 112 bytes
     [0] ./.test/declarationWatch/app.ts 67 bytes {0} [built] [1 error]
     [1] ./.test/declarationWatch/dep.ts 45 bytes {0} [built] [1 error]