… test/declarationWatch/expectedOutput/patch0/bundle.js
… test/declarationWatch/expectedOutput/patch1/bundle.js