@use-strict 2015-11-06
Very nice. Good to have this stuff documented.