@blakeembrey 2015-11-06
LGTM :+1: 

I've also seen the perf time issues be around 20% slower with `ts-loader` than `typescript-simple-loader`, but I haven't looked into it 100% yet. Might be something worth profiling against too, since other loaders haven't implemented these things but seem to still be a bit quicker.