@jbrantly 2015-11-07
Thanks!

> Might be something worth profiling against too, since other loaders haven't implemented these things but seem to still be a bit quicker.

True. Off the top of my head I bet that's (at least partially) due to `program.getDiagnostics` vs getting diagnostics for each individual file. Something to look into.